selflink/requirements/base.txt
Django==5.1.1
djangorestframework==3.15.2
django-environ==0.11.2
drf-spectacular==0.27.2
psycopg[binary]==3.2.3
